url_short/requirements.txt
Django==4.0.5
djangorestframework
PyJWT
pyshorteners
# django-cors-headers
# Pillow==9.1.1